backend/requirements.txt
Django~=5.2.8
gunicorn
python-dotenv
django-filter
markdown
djangorestframework
django-rest-swagger
setuptools
psycopg2-binary~=2.9
drf-yasg